added optimize: reuse srer out ablate

+++ exp_full.py
@@ -108,7 +108,7 @@ if __name__ == "__main__":
     parser.add_argument("--loc", type=str, default="providence", choices=["providence", "boston", "auckland", "san_francisco"], help="env name.")
     parser.add_argument("--ablate", type=str, default="text", choices=["text", "image", None], help="ablate out a modality or None to use both.")
     parser.add_argument("--nsamples", type=int, default=3, help="number of sample utts per LTL formula or None for all")
-    parser.add_argument("--seed", type=int, default=111, help="seed to random sampler.")  # 0, 1, 2, 42, 111
+    parser.add_argument("--seed", type=int, default=111, help="seed to random sampler.")  # 0, 1, 2, 42, 111 (resreved for ablate)
     parser.add_argument("--topk", type=int, default=10, help="top k most likely landmarks grounded by REG.")
     args = parser.parse_args()
     loc_id = f"{args.loc}_n{args.nsamples}_seed{args.seed}" if args.nsamples else f"{args.loc}_all_seed{args.seed}"
@@ -137,14 +137,18 @@ if __name__ == "__main__":
                         ]
     )
     logging.info(f"***** Full System Evaluation Ablate {args.ablate}: {loc_id}" if args.ablate else f"***** Full System Evaluation: {loc_id}")
+    logging.info(f"{graph_dpath}\n{osm_fpath}\n{utts_fpath}\n{true_results_fpath}\n{results_dpath}")
 
     # Spatial Referring Expression Recognition (SRER)
     srer_out_fpath_modular = os.path.join(os.path.expanduser("~"), "ground", f"results_modular_ablate_{args.ablate}" if args.ablate else "results_modular", loc_id, srer_out_fname)
-    srer_out_fpath_ablate = os.path.join(os.path.expanduser("~"), "ground", "results_full_ablate_image" if args.ablate == "text" else "results_full_ablate_text", loc_id, srer_out_fname)
-    if os.path.isfile(srer_out_fpath_modular):  # same SRER output for exp_full and exp_modular
+    srer_out_fpath_ablate_txt = os.path.join(os.path.expanduser("~"), "ground", "results_full_ablate_text", loc_id, srer_out_fname)
+    srer_out_fpath_ablate_img = os.path.join(os.path.expanduser("~"), "ground", "results_full_ablate_image", loc_id, srer_out_fname)
+    if not os.path.isfile(srer_out_fpath) and os.path.isfile(srer_out_fpath_modular):  # same SRER output for exp_full and  exp_modular
         copy2(srer_out_fpath_modular, srer_out_fpath)
-    elif os.path.isfile(srer_out_fpath_ablate):  # same SRER output for ablate text and ablate image
-        copy2(srer_out_fpath_ablate, srer_out_fpath)
+    elif not os.path.isfile(srer_out_fpath) and os.path.isfile(srer_out_fpath_ablate_txt):  # same SRER output for ablate text and ablate image
+        copy2(srer_out_fpath_ablate_txt, srer_out_fpath)
+    elif not os.path.isfile(srer_out_fpath) and os.path.isfile(srer_out_fpath_ablate_img):
+        copy2(srer_out_fpath_ablate_img, srer_out_fpath)
     else:
         run_exp_srer(utts_fpath, srer_out_fpath)
     eval_srer(true_results_fpath, srer_out_fpath)
